The all new VICE Comm manager
This has a new sustem that runs its own thread and you send it a command, and a callback for when the command runs. Known issues - the src seems to be doubled lined. The editor can get it self into odds states if you "jump over" or "jump out" and VICE never comes back.